Send Data from Lua to Radio Tower
now you can send data from FIN to Radio Tower with that function
radio:sendData('id:1234, name:herbbert, test: good')
is passed a string which is structured as follows
'key:value'

in ue4 it is pasted into a structure and pushed into an array. So an array element always consists of a key and the value. The key describes the UI element and the value is the data.